access 1

+++ App/AppStatus.py
@@ -1,9 +1,11 @@
 from django.db import models
 from django.urls import reverse  
-from App.commonpy.ComonFun import GettPostVal, UpdateError, HasUniqueAttribute, GettObjVal, CheckItemExitsInObjArray,api_responce
-#from  App.commonpy.serializers import StatusSerializer
-from django.http import JsonResponse
+import json
 
+from App.commonpy.ComonFun import GettPostVal, api_responce, UpdateError
+from .AppList import AppList
+from .AppFileds import AppFileds
+from .AppFlow import AppFlow
 
 class AppStatus(models.Model):
   id = models.AutoField(primary_key=True)
@@ -12,12 +14,12 @@ class AppStatus(models.Model):
   order = models.IntegerField(default=0)
   
   def __str__(self):
-        return str(self.id) + " : " + str(self.app_id) + " : " + self.status
+        return str(self.id) + " : " + str(self.app_id) + " : " + self.status + " : " + str(self.order)
   
   def get_absolute_url(self):  
      return reverse('AppStatus_edit', kwargs={'pk': self.pk})
   
-  def app_status_update_ajax(request):
+  def app_status_update_ajax(request):    
       err = {'msg': ''}
       id = GettPostVal(request, 'id', err, False )
       app_id = GettPostVal(request, 'app_id', err, True  )
@@ -28,7 +30,9 @@ class AppStatus(models.Model):
             if not err['msg']:
                   item = AppStatus(status = status, app_id = app_id, order = order)
                   item.save()
-                  return api_responce(item.id, 0)            
+                  return api_responce(item.id, 0)     
+            else:
+                  return api_responce(None, 1,err['msg'])       
       else:
             item = AppStatus.objects.get(id=id)
             item.status = status
@@ -40,20 +44,32 @@ class AppStatus(models.Model):
       err = {'msg': ''}
       id = GettPostVal(request, 'id', err, True )
       if id: #New item
-           item = AppStatus.objects.get(id=id)
-           item.delete()
-           return api_responce(item.id, 0)
+            valErr = {'msg': ''}
+            item = AppStatus.objects.get(id=id)
+
+            #Access
+            app_list = AppList.objects.get(id=item.app_id)
+            access = app_list.access
+            access_lst = json.loads(access)
+            for asl in access_lst: 
+                 for ac in asl['status']: 
+                      if ac == id:
+                           UpdateError(valErr, "Status is using in access " + asl['name'] )
+                           break
+                                              
+            #Flow
+            app_flows = AppFlow.objects.filter(app_id=item.app_id)
+            for flow in app_flows:
+                  if str(flow.from_status) == id:
+                        UpdateError(valErr, "Status is using in Flow From Status  for action " + flow.action )
+                        break
+                  if str(flow.to_status) == id:
+                        UpdateError(valErr, "Status is using in Flow To Status for action " + flow.action )
+                        break
+                 
+            if valErr['msg']: return api_responce(None, 1, valErr['msg'])                                             
+            item.delete()
+            return api_responce(item.id, 0)
       else:
             return api_responce(None, 1,err['msg'])
   
-#   def get_status_list_ajax(request):
-#     appid = request.POST['appid']
-#     items = AppStatus.objects.filter(app_id=appid)
-#     serializer = StatusSerializer(items, many=True)
-
-#     data = {
-#         'status': serializer.data
-#     }
-#     response = JsonResponse(data)
-#     response.status_code = 200
-#     return response
